requirements.txt
pandas
openai
transformers
datasets
nltk
spacy
scikit-learn
python-dotenv
tabulate
sentence-transformers
